Multiple bug fixes (#106)
# Description
This PR addresses the following issues:
- Closes #101 
- Closes #102 
- Closes #104 
- Closes #105 

To accomplish this, we do the following:
* For each graph update job, store a Redis hash containing all of the
transactions submitted to the chain, their birth/death blocks (for
expiration), success/failure status, and any error object
* Eliminate the `graphTxMonitorQueue`. Instead, add an additional
blockchain scanner service that walks the chain. For each block, check
against ALL transaction pending as retrieved from Redis, and update all
transactions' statuses accordingly.
* After submitting transactions to the chain, queue a child job that
monitors the pending transactions for the parent job. The parent job has
a new `state` enum in its data that indicates the state it's in. Each
time the child job runs, it will check all the candidate transactions
for its parent's transactions. If any are still pending, it adds a delay
& goes back on the queue. When none are pending (ie, all are completed,
failed, or expired), then it updates the state of the parent job with
the overall status and completes successfully. Once the child job
completes, the parent job is released to handle its final disposition.
* Moved back to using capacity batch transactions; the success event we
monitor for is `BatchCompleted`

---------

Co-authored-by: Wil Wade <wil.wade@unfinished.com>